sidebar set for the profile page
