+++ frontend/src/pages/BookAppointment.jsx
@@ -4,6 +4,7 @@ import Calendar from 'react-calendar';
 import { format } from 'date-fns';
 import 'react-calendar/dist/Calendar.css';
 import { therapists as therapistList } from './TherapistDirectory';
+import Sidebar from '../components/Sidebar';
 
 function TimeButton({ label, selected, onClick }) {
   const base = "px-4 py-2 rounded-md border text-sm";
@@ -52,15 +53,7 @@ function BookAppointment() {
   }
   return (
     <div className="flex h-screen">
-      <aside className="w-60 shrink-0 border-r border-gray-200 p-4 hidden md:block sticky top-0 h-screen overflow-y-auto">
-        <div className="font-semibold text-gray-900 mb-4">MindEase</div>
-        <nav className="space-y-2 text-sm">
-          <a className="block px-3 py-2 rounded-md hover:bg-gray-50" href="/dashboard">Dashboard</a>
-          <a className="block px-3 py-2 rounded-md bg-blue-50 text-blue-700" href="/appointments/book">Appointments</a>
-          <a className="block px-3 py-2 rounded-md hover:bg-gray-50" href="/profile">Profile</a>
-          <a className="block px-3 py-2 rounded-md hover:bg-gray-50" href="#">Settings</a>
-        </nav>
-      </aside>
+      <Sidebar/>
       <div className="flex-1 overflow-y-auto p-6 bg-gray-50">
         <h1 className="text-xl font-semibold text-gray-900 mb-4">Book Appointment</h1>
         <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
